+++ header/Sphere.h
@@ -10,8 +10,8 @@ public:
     Sphere();
     Sphere(Vec3 cen, double r, std::shared_ptr<Material> m);
 
-    virtual bool hit(const Ray& r, double t_min, double t_max, HitRecord& rec) const override;
-    virtual bool bounding_box(double time0, double time1, AABB& output_box) const override;
+    virtual bool hit(const Ray& r, float t_min, float t_max, HitRecord& rec) const override;
+    virtual bool bounding_box(float time0, float time1, AABB& output_box) const override;
 
 private:
     Vec3 center;
